+++ core/templates/base.html
@@ -98,7 +98,7 @@
                         <li class="nav-item">
                             <div class="dropdown">
                                 <a aria-expanded="false" aria-haspopup="true" class="dropdown-toggle nav-link nav-item arrow-bottom"
-                                    data-toggle="dropdown-grid" href="#" role="button">My Account</a>
+                                    data-toggle="dropdown-grid" href="#" role="button">My account</a>
                                 <div class="row dropdown-menu" style="left: -666px;">
                                     <div class="col-auto" data-dropdown-content="" style="left: 666px;">
                                         <div class="dropdown-grid-menu">
@@ -133,7 +133,7 @@
                 <div class="col-auto mt-3 mt-lg-0">
                     <ul class="list-unstyled d-flex mb-0">
                         <li class="mx-3">
-                            <a href="https://github.com/Maxamuss/code-atlas" target="_this" class="hover-fade-out">
+                            <a href="https://github.com/Maxamuss/django-notebook" target="_this" class="hover-fade-out">
                                 <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none"
                                     class="injected-svg icon icon-sm bg-white" data-src="/static/img/icons/social/github.svg/"
                                     xmlns:xlink="http://www.w3.org/1999/xlink">
